Fixed FSM to operate properly when inputs are dispatched in interrupt context.
Access the the FSM is controlled by a semaphore.  When an input is dispatched in interrupt context, the semaphore is grabbed and the input processed in the interrupt context only if the semaphore is free and sleeping is not required.  If the semaphore is not free, the input is re-dispatched via the shared workqueue, which can sleep until the semaphore is available.