Added support for the 'attacker behavior', i.e., a single machine can act as multiple identities.
The number of identities to present can be specified when loading the module via the 'numids' parameter.  The default value is 1.

Instead of maintaining a single global FSM state, all the valid FSM instances (e.g., one per active identity) are maintained in a linked list.  When packets are received, they are passed to each FSM in order.  For TXREQ packets, where response time is critical, each packet is passed only to the FSM corresponding to the requested identity.