mason: move INIT packet handling to workqueue
The wireless card cannot retrieve new TX packets while mason is processing a received packet (because a softirq cannot preempt another softirq).  Thus, the PAR packets (numbering in the hundreds when emulating an attacker) sent in response to an INIT cannot actually be sent until after the softirq creating them has finished.  So, we move this creation to a preemptable workqueue.