mason: free 'struct fsm' by ref counting, not wq flush
References to an fsm instance are held in three places: 1) the global
fsm rcu_list, 2) any dispatched inputs in the workqueue, and 3) via
container_of, in the fsm->timer.  Previously, FSMs were freed only
after all references were removed by removing from the rcu_list,
flushing the work_queue, canceling the timer.  However, this process
was done in softirq context, so if work_queue flush slept, the kernel
BUGed and crashed.

To fix this, we instead reference count the fsm instances, with the
rcu_list and dispatched workqueue inputs each incrementing the
refcount.  When the refcount drops to zero, the associated timer is
canceled and the fsm freed.

A new fsm.run member is set to zero when the fsm deletion is first
requested.  New timers and inputs are not created if the fsm.run is
zero. Thus, eventually all the existing inputs will be processed, the
refcount will drop to zero, and the fsm will be freed.