mason: add spinlock to id_table struct
The spinlock is used when updating the table contents by adding an
identity or recording a new RSSI observation.  The id_table is shared
state for attacker clients, so concurrent write attempts are possible
while in the states associated with these actions.  Other reads are
not protected because in those states, no one is writing.